Fix Curve ABI and remove hardcoded chainId fallbacks
- Update Curve pool ABI to use uint256 for token indices (newer pools)
- Remove hardcoded chainId fallbacks from all hooks
- Hooks now require valid chainId before making contract calls
- Add proper null checks and enabled conditions to prevent calls
  when wallet is not connected to a supported network

+++ src/contracts/abis/CurvePool.ts
@@ -1,12 +1,13 @@
 // Curve StableSwap Pool ABI - USDC/DXY-BEAR trading
+// Note: Using uint256 for indices (newer Curve pools)
 export const CURVE_POOL_ABI = [
   {
     type: 'function',
     name: 'exchange',
     stateMutability: 'nonpayable',
     inputs: [
-      { name: 'i', type: 'int128' },
-      { name: 'j', type: 'int128' },
+      { name: 'i', type: 'uint256' },
+      { name: 'j', type: 'uint256' },
       { name: 'dx', type: 'uint256' },
       { name: 'min_dy', type: 'uint256' },
       { name: 'receiver', type: 'address' },
@@ -18,8 +19,8 @@ export const CURVE_POOL_ABI = [
     name: 'get_dy',
     stateMutability: 'view',
     inputs: [
-      { name: 'i', type: 'int128' },
-      { name: 'j', type: 'int128' },
+      { name: 'i', type: 'uint256' },
+      { name: 'j', type: 'uint256' },
       { name: 'dx', type: 'uint256' },
     ],
     outputs: [{ type: 'uint256', name: 'dy' }],
@@ -52,7 +53,7 @@ export const CURVE_POOL_ABI = [
     stateMutability: 'nonpayable',
     inputs: [
       { name: '_token_amount', type: 'uint256' },
-      { name: 'i', type: 'int128' },
+      { name: 'i', type: 'uint256' },
       { name: '_min_amount', type: 'uint256' },
       { name: 'receiver', type: 'address' },
     ],
@@ -84,9 +85,9 @@ export const CURVE_POOL_ABI = [
     name: 'TokenExchange',
     inputs: [
       { name: 'buyer', type: 'address', indexed: true },
-      { name: 'sold_id', type: 'int128', indexed: false },
+      { name: 'sold_id', type: 'uint256', indexed: false },
       { name: 'tokens_sold', type: 'uint256', indexed: false },
-      { name: 'bought_id', type: 'int128', indexed: false },
+      { name: 'bought_id', type: 'uint256', indexed: false },
       { name: 'tokens_bought', type: 'uint256', indexed: false },
     ],
   },
